[WIP] remove server crate from IR, instead, add job_service.proto only
