+++ research/query_service/ir/Cargo.toml
@@ -5,7 +5,6 @@ members = [
     "graph_proxy",
     "integrated",
     "runtime",
-    "server"
 ]
 
 [profile.dev]
