+++ research/query_service/ir/clients/rust/client/Cargo.toml
@@ -6,10 +6,15 @@ edition = "2018"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-runtime_server = { path = "../../../server" }
+#runtime_server = { path = "../../../server" }
 pegasus_server = { path = "../../../../../engine/pegasus/server" }
 pegasus = { path = "../../../../../engine/pegasus/pegasus" }
 tonic = "0.6"
 prost = "0.9"
 tokio = { version = "1.0", features = ["macros", "sync", "rt-multi-thread"] }
 tokio-stream = { version = "0.1.3", features = ["net"] }
+ir_common = {path = "../../../common"}
+
+[features]
+default = []
+proto_inplace = ["ir_common/proto_inplace", "pegasus_server/gcip"]
