+++ research/query_service/ir/common/build.rs
@@ -29,6 +29,7 @@ fn codegen_inplace() -> Result<(), Box<dyn std::error::Error>> {
     println!("cargo:rerun-if-changed=../proto/algebra.proto");
     println!("cargo:rerun-if-changed=../proto/schema.proto");
     println!("cargo:rerun-if-changed=../proto/results.proto");
+    println!("cargo:rerun-if-changed=../proto/job_service.proto");
     let out_dir = PathBuf::from(GEN_DIR);
     if out_dir.exists() {
         let _ = std::fs::remove_dir_all(GEN_DIR);
@@ -44,6 +45,7 @@ fn codegen_inplace() -> Result<(), Box<dyn std::error::Error>> {
                 "../proto/algebra.proto",
                 "../proto/schema.proto",
                 "../proto/results.proto",
+                "../proto/job_service.proto",
             ],
             &["../proto"],
         )?;
@@ -62,6 +64,7 @@ fn codegen_inplace() -> Result<(), Box<dyn std::error::Error>> {
                 "../proto/algebra.proto",
                 "../proto/schema.proto",
                 "../proto/results.proto",
+                "../proto/job_service.proto",
             ],
             &["../proto"],
         )?;
