+++ research/query_service/ir/core/Cargo.toml
@@ -14,7 +14,7 @@ lazy_static = "1.3.0"
 log = "0.4"
 #pegasus_client = {path = "../../../engine/pegasus/clients/rust/client"}
 pegasus_client = {path = "../clients/rust/client"}
-runtime_server = {path = "../server"}
+#runtime_server = {path = "../server"}
 pegasus = {path = "../../../engine/pegasus/pegasus"}
 prost = "0.9"
 serde = "1.0"
