+++ research/query_service/ir/integrated/Cargo.toml
@@ -16,7 +16,7 @@ clap = "2.32.0"
 pegasus = { path = "../../../engine/pegasus/pegasus" }
 pegasus_client = {path = "../clients/rust/client"}
 pegasus_server = { path = "../../../engine/pegasus/server" }
-runtime_server = {path = "../server"}
+#runtime_server = {path = "../server"}
 pegasus_common = { path = "../../../engine/pegasus/common" }
 ir_common = {path = "../common"}
 ir_core = {path = "../core"}
