+++ research/query_service/ir/runtime/Cargo.toml
@@ -12,7 +12,7 @@ log = "0.4"
 pegasus_common = { path = "../../../engine/pegasus/common" }
 pegasus = { path = "../../../engine/pegasus/pegasus" }
 pegasus_server = { path = "../../../engine/pegasus/server" }
-runtime_server = {path = "../server"}
+#runtime_server = {path = "../server"}
 prost = "0.9"
 vec_map = "0.8.2"
