--- research/query_service/ir/server/Cargo.toml
@@ -1,33 +1,0 @@
-[package]
-name = "runtime_server"
-version = "0.1.0"
-authors = ["chenqiang.mcq <chenqiang.mcq@alibaba-inc.com>"]
-edition = "2018"
-
-[dependencies]
-pegasus_common = { path = "../../../engine/pegasus/common" }
-pegasus_network = { path = "../../../engine/pegasus/network"}
-pegasus_memory = { path = "../../../engine/pegasus/memory"}
-pegasus = { path = "../../../engine/pegasus/pegasus" }
-log = "0.4"
-crossbeam-utils = "0.6"
-#crossbeam-channel = "0.3.6"
-tonic = "0.6"
-prost = "0.9"
-tokio = { version = "1.0", features = ["macros", "sync", "rt-multi-thread"] }
-tokio-stream = { version = "0.1.3", features = ["net"] }
-toml = "0.5"
-serde = { version = "1.0", features = ["derive"] }
-structopt = "0.2"
-
-[dev-dependencies]
-libloading = "0.7"
-
-[build-dependencies]
-tonic-build = "0.5"
-
-[features]
-default = []
-# set to generate code in place(generated codes are in current codebase);
-gcip = []
-
